requirements.txt
Flask
Flask-Cors
Flask-WTF
google-cloud-aiplatform
mysql-connector-python
bcrypt